Updated regEx call in ctor to handle rotated file naming scheme used in our code.
Updated create method to call checkRetentionSync() method instead of calling the async version.

Added checkRetentionSync() method to this module.  Retention checking and purging of old files can result in exceptions when this module is consumed through multiple consumers.  The async version of this routine does not handle situations where multiple directory reads results in the same block of files being processed asynchronously, and file unlink calls occur multiple times on the same file.  This results in an exception being thrown and process termination."